refactor(rest): move slugify, catalog doc validators and _ADA_META to ada.core
ada.cadit.dexpi imported slugify/validate_equipment_doc from
ada.comms.rest.catalog (5 sites) and ada.topo_model.excel_io imported the
ADA_META_* constants from ada.comms.rest.procedural: modelling code depending
on the REST layer. Both helpers are stdlib/pydantic-only, so they now live in
ada.core (text_utils, catalog_docs, xlsx_meta) below every consumer. The old
modules re-export them, so out-of-tree imports keep working.

ada.topo_model was not an option for the neutral home: importing any of its
submodules runs its package __init__ and drags the modelling stack into the
slim viewer runtime (tests/comms/test_engine_specs_slim_runtime.py). The
Dockerfiles ship the three new modules alongside ada.core.file_system.
