+++ deploy/Dockerfile.viewer-fast
@@ -55,6 +55,13 @@ COPY src/ada/cad/ /app/src/ada/cad/
 # See the matching COPY in Dockerfile.viewer for why only this one module ships.
 COPY src/ada/core/__init__.py /app/src/ada/core/__init__.py
 COPY src/ada/core/file_system.py /app/src/ada/core/file_system.py
+# ada.core.{text_utils,catalog_docs,xlsx_meta} — the slug helper, the equipment/system doc
+# validators and the _ADA_META sheet reader that ada.comms.rest.{catalog,procedural} re-export.
+# Stdlib + pydantic only, like file_system.py above: they moved here so the CAD readers and the
+# procedural engines stop importing from ada.comms.
+COPY src/ada/core/text_utils.py /app/src/ada/core/text_utils.py
+COPY src/ada/core/catalog_docs.py /app/src/ada/core/catalog_docs.py
+COPY src/ada/core/xlsx_meta.py /app/src/ada/core/xlsx_meta.py
 COPY src/ada/sections/categories.py        /app/src/ada/sections/categories.py
 COPY src/ada/sections/profile_lookup.py    /app/src/ada/sections/profile_lookup.py
 COPY src/ada/sections/resources/ProfileDB.json /app/src/ada/sections/resources/ProfileDB.json
